refactor git interactions
