+++ src/run.ts
@@ -1,65 +1,27 @@
-import { exec, getExecOutput } from "@actions/exec";
-import { GitHub, getOctokitOptions } from "@actions/github/lib/utils";
-import * as github from "@actions/github";
 import * as core from "@actions/core";
-import fs from "fs-extra";
-import { getPackages, Package } from "@manypkg/get-packages";
-import path from "path";
-import * as semver from "semver";
+import { exec, getExecOutput } from "@actions/exec";
+import * as github from "@actions/github";
 import { PreState } from "@changesets/types";
-import {
-  getChangelogEntry,
-  getChangedPackages,
-  sortTheThings,
-  getVersionsByDirectory,
-} from "./utils";
-import * as gitUtils from "./gitUtils";
-import readChangesetState from "./readChangesetState";
+import { Package, getPackages } from "@manypkg/get-packages";
+import fs from "fs-extra";
+import path from "path";
 import resolveFrom from "resolve-from";
-import { throttling } from "@octokit/plugin-throttling";
-import { commitChangesFromRepo } from "@s0/ghcommit/git";
+import * as semver from "semver";
+import { Git } from "./git";
+import { Octokit } from "./octokit";
+import readChangesetState from "./readChangesetState";
+import {
+  getChangedPackages,
+  getChangelogEntry,
+  getVersionsByDirectory,
+  sortTheThings,
+} from "./utils";
 
 // GitHub Issues/PRs messages have a max size limit on the
 // message body payload.
 // `body is too long (maximum is 65536 characters)`.
 // To avoid that, we ensure to cap the message to 60k chars.
 const MAX_CHARACTERS_PER_MESSAGE = 60000;
-
-export const setupOctokit = (githubToken: string) => {
-  return new (GitHub.plugin(throttling))(
-    getOctokitOptions(githubToken, {
-      throttle: {
-        onRateLimit: (retryAfter, options: any, octokit, retryCount) => {
-          core.warning(
-            `Request quota exhausted for request ${options.method} ${options.url}`
-          );
-
-          if (retryCount <= 2) {
-            core.info(`Retrying after ${retryAfter} seconds!`);
-            return true;
-          }
-        },
-        onSecondaryRateLimit: (
-          retryAfter,
-          options: any,
-          octokit,
-          retryCount
-        ) => {
-          core.warning(
-            `SecondaryRateLimit detected for request ${options.method} ${options.url}`
-          );
-
-          if (retryCount <= 2) {
-            core.info(`Retrying after ${retryAfter} seconds!`);
-            return true;
-          }
-        },
-      },
-    })
-  );
-};
-
-export type Octokit = ReturnType<typeof setupOctokit>;
 
 const createRelease = async (
   octokit: Octokit,
@@ -99,45 +61,11 @@ const createRelease = async (
   }
 };
 
-export type GitTaggingStrategy = {
-  pushAllTags: () => Promise<void>;
-  pushTag: (tag: string) => Promise<void>;
-};
-
-export const getCliTaggingStrategy = (): GitTaggingStrategy => ({
-  pushAllTags: async () => {
-    await gitUtils.pushTags();
-  },
-  pushTag: async () => {
-    // When using the CLI, we push all tags together at once
-  },
-});
-
-export const getApiTaggingStrategy = (
-  octokit: Octokit
-): GitTaggingStrategy => ({
-  pushAllTags: async () => {
-    // When using the API, we push all tags individually
-  },
-  pushTag: async (tagName: string) => {
-    await octokit.rest.git
-      .createRef({
-        ...github.context.repo,
-        ref: `refs/tags/${tagName}`,
-        sha: github.context.sha,
-      })
-      .catch((err) => {
-        // Assuming tag was manually pushed in custom publish script
-        core.warning(`Failed to create tag ${tagName}: ${err.message}`);
-      });
-  },
-});
-
 type PublishOptions = {
   script: string;
   octokit: Octokit;
   createGithubReleases: boolean;
-  taggingStrategy: GitTaggingStrategy;
+  git: Git;
   cwd?: string;
 };
 
@@ -154,9 +82,9 @@ type PublishResult =
 
 export async function runPublish({
   script,
+  git,
   octokit,
   createGithubReleases,
-  taggingStrategy,
   cwd = process.cwd(),
 }: PublishOptions): Promise<PublishResult> {
   let [publishCommand, ...publishArgs] = script.split(/\s+/);
@@ -166,8 +94,6 @@ export async function runPublish({
     publishArgs,
     { cwd }
   );
-
-  await taggingStrategy.pushAllTags();
 
   let { packages, tool } = await getPackages(cwd);
   let releasedPackages: Package[] = [];
@@ -196,7 +122,7 @@ export async function runPublish({
       await Promise.all(
         releasedPackages.map(async (pkg) => {
           const tagName = `${pkg.packageJson.name}@${pkg.packageJson.version}`;
-          await taggingStrategy.pushTag(tagName);
+          await git.pushTag(tagName);
           await createRelease(octokit, { pkg, tagName });
         })
       );
@@ -218,7 +144,7 @@ export async function runPublish({
         releasedPackages.push(pkg);
         if (createGithubReleases) {
           const tagName = `v${pkg.packageJson.version}`;
-          await taggingStrategy.pushTag(tagName);
+          await git.pushTag(tagName);
           await createRelease(octokit, { pkg, tagName });
         }
         break;
@@ -326,54 +252,15 @@ export async function getVersionPrBody({
   return fullMessage;
 }
 
-type GitPushStrategy = {
-  prepareVersionBranch: (versionBranch: string) => Promise<void>;
-  pushChanges: (params: {
-    versionBranch: string;
-    finalCommitMessage: string;
-  }) => Promise<void>;
-};
-
-export const getCliPushStrategy = (): GitPushStrategy => ({
-  prepareVersionBranch: async (versionBranch: string) => {
-    await gitUtils.switchToMaybeExistingBranch(versionBranch);
-    await gitUtils.reset(github.context.sha);
-  },
-  pushChanges: async ({ versionBranch, finalCommitMessage }) => {
-    if (!(await gitUtils.checkIfClean())) {
-      await gitUtils.commitAll(finalCommitMessage);
-    }
-    await gitUtils.push(versionBranch, { force: true });
-  },
-});
-
-export const getApiPushStrategy = (octokit: Octokit): GitPushStrategy => ({
-  prepareVersionBranch: async () => {
-    // Preparing a new local branch is not necessary when using the API
-  },
-  pushChanges: async ({ versionBranch, finalCommitMessage }) => {
-    await commitChangesFromRepo({
-      octokit,
-      ...github.context.repo,
-      branch: versionBranch,
-      message: finalCommitMessage,
-      base: {
-        commit: github.context.sha,
-      },
-      force: true,
-    });
-  },
-});
-
 type VersionOptions = {
   script?: string;
+  git: Git;
   octokit: Octokit;
   cwd?: string;
   prTitle?: string;
   commitMessage?: string;
   hasPublishScript?: boolean;
   prBodyMaxCharacters?: number;
-  gitPushStrategy: GitPushStrategy;
   branch?: string;
 };
 
@@ -383,13 +270,13 @@ type RunVersionResult = {
 
 export async function runVersion({
   script,
+  git,
   octokit,
   cwd = process.cwd(),
   prTitle = "Version Packages",
   commitMessage = "Version Packages",
   hasPublishScript = false,
   prBodyMaxCharacters = MAX_CHARACTERS_PER_MESSAGE,
-  gitPushStrategy,
   branch,
 }: VersionOptions): Promise<RunVersionResult> {
   let repo = `${github.context.repo.owner}/${github.context.repo.repo}`;
@@ -398,7 +285,7 @@ export async function runVersion({
 
   let { preState } = await readChangesetState(cwd);
 
-  await gitPushStrategy.prepareVersionBranch(versionBranch);
+  await git.prepareBranch(versionBranch);
 
   let versionsByDirectory = await getVersionsByDirectory(cwd);
 
@@ -444,7 +331,7 @@ export async function runVersion({
     !!preState ? ` (${preState.tag})` : ""
   }`;
 
-  await gitPushStrategy.pushChanges({ versionBranch, finalCommitMessage });
+  await git.pushChanges({ branch: versionBranch, message: finalCommitMessage });
 
   let existingPullRequests = await existingPullRequestsPromise;
   core.info(JSON.stringify(existingPullRequests.data, null, 2));
